crawlers/requirements.txt
# AI Frontier Watch - Crawler Dependencies

# Core dependencies
requests>=2.28.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
lxml>=4.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Redis for production retry queue
# redis>=5.0.0

# Optional: Supabase client for production storage
# supabase>=2.0.0
